wl1744 - fix compiler problems
ndb/include/debugger/GrepError.hpp:
  prexif
ndb/include/kernel/LogLevel.hpp:
  fix correct type prototype
ndb/include/kernel/signaldata/FsCloseReq.hpp:
  remove usage of true
ndb/include/mgmapi/mgmapi.h:
  use Uint64 instead of long long
ndb/include/mgmcommon/IPCConfig.hpp:
  correct type
ndb/include/portlib/NdbTCP.h:
  fix #elif
ndb/include/transporter/TransporterRegistry.hpp:
  correct type
ndb/include/util/Parser.hpp:
  correct type
  made stuff public as vc++ couldn't handle template friends
ndb/include/util/SimpleProperties.hpp:
  correct type
ndb/src/common/debugger/EventLogger.cpp:
  Prefix GrepError::Code as GrepError::GE_Code
ndb/src/common/debugger/GrepError.cpp:
  Prefix GrepError::Code as GrepError::GE_Code
ndb/src/common/debugger/signaldata/FsCloseReq.cpp:
  removed usage of true
ndb/src/common/debugger/signaldata/FsReadWriteReq.cpp:
  removed usage of true
ndb/src/common/logger/LogHandlerList.hpp:
  use ndb_global
ndb/src/common/mgmcommon/ConfigRetriever.cpp:
  removed ConfigRetriever::get_config(file) from windows
ndb/src/common/transporter/TransporterRegistry.cpp:
  interface is a reserved word in vc++
ndb/src/kernel/blocks/grep/Grep.hpp:
  Prefix GrepError::Code as GrepError::GE_Code
ndb/src/kernel/vm/Configuration.cpp:
  Use BaseString::snprintf
ndb/src/kernel/vm/Configuration.hpp:
  correct type
ndb/src/kernel/vm/SimplePropertiesSection.cpp:
  correct type
ndb/src/mgmapi/mgmapi_configuration.cpp:
  use Uin6t4
ndb/src/mgmapi/mgmapi_configuration.hpp:
  use Uin6t4
ndb/src/mgmsrv/ConfigInfo.cpp:
  Prefix
ndb/src/mgmsrv/ConfigInfo.hpp:
  prefix
ndb/src/ndbapi/Ndb.cpp:
  removed usued include
ndb/src/ndbapi/NdbImpl.hpp:
  prefix
ndb/src/ndbapi/ObjectMap.hpp:
  Better typecast

+++ ndb/include/mgmapi/mgmapi.h
@@ -730,7 +730,7 @@ extern "C" {
   int ndb_mgm_get_int_parameter(const ndb_mgm_configuration_iterator*, 
 				int param, unsigned * value);
   int ndb_mgm_get_int64_parameter(const ndb_mgm_configuration_iterator*,
-				  int param, unsigned long long * value);
+				  int param, Uint64 * value);
   int ndb_mgm_get_string_parameter(const ndb_mgm_configuration_iterator*,
 				   int param, const char  ** value);
 #ifdef __cplusplus
